Titanic: first neural nets
This commit contains code2.py, which is just code.py with logistic
regression swapped for neural nets, and some mean normalisation of the
training data. It also contains the submission2.csv file.

Additionally, it also has a .gitignore file, to hide the Intellij IDEA
configuration files, since I'm now using that IDE to help with coding.