Titanic: code4e.py and output files
This commit contains code4e.py, which is the same as code4d.py, except
that it removes the manual cross validation code, and only trains
models for 500 iterations, instead of 100, 200, 300, 400, and 500
iterations.

The results were grossly disappointing: Saga: 0.56459, Sag: 0.56459,
Lbfgs: 0.56937, Lib-Linear: 0.53110, Newton-CG: 0.55980.

The files submitted to Kaggle are included in this commit.