Titanic: code5a.py First use of PCA with LR
code5a.py contains the first use of PCA prior to logistic regression.

There is no cross validation, and scoring is only carried out on the
training data, so the results that this code outputs to the command
line shouldn't be taken as particularly indicitive of the quality of
this approach.

The script will now be improved to incorporate manual cross validation
as a way of verifying the n_components parameter, prior to the best
option being selected to use for generating a submission file for
Kagggle.