HousePrices: Beginning code
This commit contains 3 code files; for the exploration and
understanding of the data, the training of models on the data, and for
the generation of submittable files for the Kaggle competition.

It also contains a readme.md file, so that information relevant to the
understanding of the code can be put there.