HousePrices: add MLPClassifier and SVM.SVC
This adds the the neural network classifier, in the form of an
MLPClassifier (it has very poor performance), and an SVM classifier,
in the form of a SVC classifier (which also has very poor performance).

Additionally, due to the variable performance of linear regression
depending on the selection of training and test sets, the process is
put into a loop which exits on the linear regression classifier getting
a score of 0.88 or above.