HousePrices: Modify preprocessing for submitting
This modifies the preprocessing so that the test.csv data has the same
columns as the train.csv data does once the preprocessing is done.

This is done because pd.get_dummies() results in different results due
to test.csv features not containing all the same variations of data as
train.csv features.