Digits: COMPLETE
This commit contains the entire digis analysis. I is the first I have
done that identifies different images, which makes it particularly
exciting. I hope to do more advanced and 'cool' types of ML and AI
stuff going forward.

Also note that critical errors in the code for the Iris project which
were identified during this one, and I will go back and fix them before
starting on the next one.