Titanic2: fixing an even more critical error.
I included 'Survived' in the training data - so no wonder the
predictions were all perfect.

I'll just rewrite the report according to the new results. It would be
boring and uninstructive (useless, in other words) for anyone to have to
look throught all of the work so far, to then have to re-read a full
re-write.